requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
streamlit>=1.24.0
google-generativeai>=0.3.2
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.31.0
python-dotenv>=1.0.0